add component for fetch api and gitignore more files
