+++ .gitignore
@@ -22,3 +22,11 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+
+#TXT
+ignore.txt
+
+#JSON
+notice.json
+notice-img.json
+notice-self.json
