+++ src/App.jsx
@@ -1,9 +1,13 @@
 import './App.css'
+import DetailsOfWanted from './components/API/Details/Details';
 import WorldMap from './components/Map/WorldMap/WorldMap'
 
 function App() {
   return (
+    <>
     <WorldMap />
+    <DetailsOfWanted />
+    </>
   );
 }
